Tweaks for troubleshooting app crashing.
